Add files via upload
The first working version.
The previous version, which did not work, attempted to find the Close button of a dialog box which appeared after clicking the Dismiss button. (It could be DOM change by Javascript.)
Windsurf tried several times but failed.

This version, instead, updates the whole page after clicking the DIsmiss button.  It does not try to find the Close button.  It re-opens the same page and checks if there is more Dismiss button.

Dismiss button 暫緩
Close button 關閉